% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ConvexHullMD}
\alias{ConvexHullMD}
\title{ConvexHullMD}
\usage{
ConvexHullMD(LofV)
}
\arguments{
\item{LofV}{List of 2D vectors.}
}
\value{
List of 2D vectors representing vertices of the convex hull.
}
\description{
Uses Jarvis March algorithm to find the convex hull of a list of 2D vectors.
}
