% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HullPlotMD.R
\name{HullPlotMD}
\alias{HullPlotMD}
\title{HullPlotMD}
\usage{
HullPlotMD(LofV)
}
\arguments{
\item{LofV}{List of 2D vectors.}
}
\value{
A plot of the data and the convex hull.
}
\description{
Uses Jarvis March to find and plot the convex hull of a list of 2D vectors.
}
